Simplified code, added gitIgnore file, removed PokerHand and verified GetBestHand.
