--- CardFace.cs
@@ -1,20 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Collections;
-using System.Text;
-using static PokerGame.Card;
-using static PokerGame.Dealer;
-
-namespace PokerGame
-{
-    public enum CardFace
-    {
-         Two, Three, Four, Five, Six, Seven, Eight, Nine, Ten, Jack, Queen, King, Ace
-    }
-
-
-}
-        
-
-
-
