--- CardSuit.cs
@@ -1,17 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Collections;
-using System.Text;
-
-namespace PokerGame
-{
-    public enum CardSuit
-    {
-        Clubs, Diamonds, Hearts, Spades
-    }
-
-}
-        
-
-
-
