+++ Decision.cs
@@ -1,23 +1,15 @@
-﻿using System;
+﻿using PokerGame.Enums;
+
 namespace PokerGame
 {
-    public class Decision
+    public partial class Decision
     {
         public DecisionType SelDecisionType { get; }
         public double Amount { get; }
-
-        public enum DecisionType
-        {
-            Fold = 1,
-            Call,
-            Raise
-        }
 
         public Decision(DecisionType selDecisionType)
         {
             SelDecisionType = selDecisionType;            
         }
     }
-}   
-
-
+}
