+++ Enums/CardFace.cs
@@ -1,0 +1,7 @@
+namespace PokerGame.Enums
+{
+    public enum CardFace
+    {
+         Two, Three, Four, Five, Six, Seven, Eight, Nine, Ten, Jack, Queen, King, Ace
+    }
+}
