+++ Enums/CardSuit.cs
@@ -1,0 +1,7 @@
+namespace PokerGame.Enums
+{
+    public enum CardSuit
+    {
+        Clubs, Diamonds, Hearts, Spades
+    }
+}
