+++ Enums/DecisionType.cs
@@ -1,0 +1,9 @@
+﻿namespace PokerGame.Enums
+{
+    public enum DecisionType
+    {
+        Fold = 1,
+        Call,
+        Raise
+    }
+}
