+++ Enums/WinningHands.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace PokerGame
+﻿namespace PokerGame.Enums
 {
     //Not critical for the code currently. Was going to create an enum of cards 
     public enum WinningHands
