+++ LoanShark.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections;
-using System.Text;
-using static PokerGame.HumanPlayer;
 
 namespace PokerGame
 {
