+++ Main.cs
@@ -1,15 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Collections;
-using System.Text;
-using static PokerGame.Card;
-using static PokerGame.Round;
-using static PokerGame.Dealer;
-using static PokerGame.CardSuit;
-using static PokerGame.CardFace;
-using static PokerGame.HumanPlayer;
-using PokerGame.Extensions;
-using static PokerGame.Game;
 
 namespace PokerGame
 {
